refactor: replace ArrayBuilder with new SegmentedArrayBuilder for improved memory management and performance.

+++ NetFabric.Hyperlinq/Extensions/Operations/Filtering/Enumerables/WhereReadOnlySpanEnumerable.cs
@@ -42,7 +42,7 @@ public readonly ref struct WhereReadOnlySpanEnumerable<TSource, TPredicate>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool MoveNext()
         {
-            while (++index < source.Length)
+            while ((uint)++index < (uint)source.Length)
             {
                 if (predicate.Invoke(source[index]))
                 {
